pyproject.toml
[tool.poetry]
name = "repl_python3_DSSystemsLabs"
version = "0.1.0"
description = ""
authors = ["NikitaSaxena1 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
pandas = "^1.4.2"
schedule = "^1.1.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
